Don't be as restrictive about peer dependencies. (#2)
This updates the way we declare our peer dependencies so that
we merely ask that the user have *at least* the provided version
installed.

While there's some risk that there will be breaking changes in the
versions we allow, I think this is the right trade-off.

We don't often update this library, or assess it's compatibility
with the latest n' greatest. This means that a security-concious
user will be harassed by `yarn` or `npm` when being diligent and
updating their dependencies (if they happen to update one expresed
as a peer dependency here).

We've also seen in practice that this is likely ok, as we *have*
updated the depedencies this module relies upon and things have
continued to work without issue.

So, by doing this we *may* make an error when there is a breaking
change in one of these dependencies harder to catch -- but this is
likely a rare occurrence and in the meantime we'll encourage rather
than discourage updates.